relocate cookie class to live alongside headers
