+++ lib/secure_headers.rb
@@ -1,5 +1,5 @@
 require "secure_headers/configuration"
-require "secure_headers/cookie"
+require "secure_headers/headers/cookie"
 require "secure_headers/headers/public_key_pins"
 require "secure_headers/headers/content_security_policy"
 require "secure_headers/headers/x_frame_options"
