+++ lib/secure_headers/headers/cookie.rb
